Moved flow steps, step sequence and step links out of the main flow file
