+++ rasa/dialogue_understanding/patterns/correction.py
@@ -7,7 +7,7 @@ from rasa.dialogue_understanding.patterns.collect_information import (
 )
 from rasa.dialogue_understanding.stack.dialogue_stack import DialogueStack
 from rasa.shared.constants import RASA_DEFAULT_FLOW_PATTERN_PREFIX
-from rasa.shared.core.flows.flow import (
+from rasa.shared.core.flows.flow_step import (
     START_STEP,
 )
 from rasa.shared.core.trackers import (
@@ -30,7 +30,7 @@ from rasa.shared.core.events import (
     SlotSet,
 )
 from rasa.core.nlg import NaturalLanguageGenerator
-from rasa.shared.core.flows.flow import END_STEP, ContinueFlowStep
+from rasa.shared.core.flows.flow_step import END_STEP, ContinueFlowStep
 
 structlogger = structlog.get_logger()
