+++ rasa/dialogue_understanding/stack/utils.py
@@ -5,7 +5,7 @@ from rasa.dialogue_understanding.patterns.collect_information import (
 from rasa.dialogue_understanding.stack.frames import BaseFlowStackFrame
 from rasa.dialogue_understanding.stack.dialogue_stack import DialogueStack
 from rasa.dialogue_understanding.stack.frames import UserFlowStackFrame
-from rasa.shared.core.flows.flow import END_STEP, ContinueFlowStep
+from rasa.shared.core.flows.flow_step import END_STEP, ContinueFlowStep
 from rasa.shared.core.flows.flows_list import FlowsList
