+++ tests/dialogue_understanding/generator/test_llm_command_generator.py
@@ -26,7 +26,7 @@ from rasa.engine.storage.local_model_storage import LocalModelStorage
 from rasa.engine.storage.resource import Resource
 from rasa.engine.storage.storage import ModelStorage
 from rasa.shared.core.events import BotUttered, SlotSet, UserUttered
-from rasa.shared.core.flows.flow import (
+from rasa.shared.core.flows.flow_step import (
     CollectInformationFlowStep,
     SlotRejection,
 )
